chore: move gix mod into git mod
